Guard mIsPendingCellInfoRequest with synch block
In the event of a modem timeout, there is an instance of
mIsPendingCellInfoRequest that isn't guarded by a sync
block. Due to the behavior of the system, I'm unable to
figure out whether this is actually a problem, though it's
theoretically possible to have it drop a response in the
event that the cache doesn't sync between the time of a
request and a timeout. Becuase it's safe to guard, and
because it's documented to be guarded by
mPendingCellInfoRequests, adding a sync block for safety.

Bug: 145965248
Test: manual

CRs-Fixed: 2722957
Change-Id: I1ca570011c8689a55b8bbbd39cda2e17338a5ac7
(cherry picked from commit 525fa6de2d8696734bc9be8fbdceda3c0932abe1)
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: ghou8s <ghou8s@gmail.com>